Refactored Ring Color Information
Previously RingData used to store a Color value instead which will cause problem with development when trying to compare Colors directly.

Now RingData stores ColorIndexs using an enum which is RingBehaviours grabs the color information from a list of registered colors in ColorManager. This way, it will be easier to compare rings against each other and also easier to change color construction data in the future.